requirements.txt
annotated-types==0.6.0
certifi==2024.2.2
charset-normalizer==3.3.2
httplib2==0.22.0
idna==3.6
inflection==0.5.1
oauth2client==4.1.3
pyairtable==2.3.0.post1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.3
pydantic_core==2.16.3
pyparsing==3.1.1
pyspark==3.1.2
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
typing_extensions==4.10.0
urllib3==2.2.1
